Observe upstream DNS files every minute
Doing the same thing in only `scheduler upstream` is not sufficient
because the unrestricted upstream file (adguard_upstream) is not
allowed to be created when the task has not been accomplished, even if
`scheduler upstream allow` is called. However, it is weird to always
create restricted file (superman_upstream) without counting the number
of upstream files just to avoid the problem. Checking these types of
things should be done by a dedicated script, which was created in this
commit.