Do not require todo_work --unchecked 0 for YouTube Music
I wanted to listen to the music after a shower today, but I didn't because retrospective tasks weren't been finished right after coming back home and taking a shower. So I decided to deregulate the condition for YouTube Music.

I know the conditions between todo_work qtg0000 and qtg1000 are exact the same, but left it as is for the possible future revert.